Modular MVVP switch part 2
Almost done with the switch.
Moved global variable from the data module to the OpenModule
Renamed OpenModule To global Module
Added every new script or modified script to  GameObject in unity
Remove RewardMenu and made reward and reward UI generic by only using text and images
Changed choices class to have a name for every choice and an index corresponding to the index they had in the menu. This allow to find what they are replacing no matter how it's handle afterward